clustermesh-apiserver: move to a subcommand
Move the current clustermesh-apiserver code to a dedicated subcommand,
to prepare for the subsequent merging of kvstoremesh into the same binary.
